pyproject.toml
[project]
name = "GitprofileBuilder"
version = "0.1.0"
description = "This is a project to create a git profile based out of resume"
authors = [
    {name = "Subhajit Pal", email = "subhajitpal7@gmail.com"},
]
dependencies = [
    "langchain>=0.2.5",
    "langchain-community>=0.2.5",
    "pypdfium2>=4.30.0",
    "faiss-cpu>=1.8.0.post1",
    "sentence-transformers>=3.1.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.0",
    "langchain-google-genai>=2.0.7",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true
